+++ tagpack/cassandra.py
@@ -1,6 +1,4 @@
-"""Cassandra - Handles ingest into cassandra"""
-import json
-import re
+"""Cassandra - Handles ingest into Apache Cassandra"""
 
 import importlib.resources as pkg_resources
 
@@ -8,32 +6,18 @@ from cassandra.cluster import Cluster
 from cassandra.concurrent import execute_concurrent
 
 from . import db
+from tagpack import StorageError
+from tagpack.tagpack import AddressTag, EntityTag
 
 CONCURRENCY = 100
 DEFAULT_TIMEOUT = 60
-
-LABEL_NORM_PATTERN = re.compile(r'[\W_]+', re.UNICODE)
 
 SCHEMA_FILE = 'tagpack_schema.cql'
 KEYPACE_PACEHOLDER = 'KEYSPACE_NAME'
 
 
-class StorageError(Exception):
-    """Class for Cassandra-related errors"""
-
-    def __init__(self, message, nested_exception=None):
-        super().__init__("Cassandra Error: " + message)
-        self.nested_exception = nested_exception
-
-    def __str__(self):
-        msg = super(StorageError, self).__str__()
-        if self.nested_exception:
-            msg = msg + '\nError Details: ' + str(self.nested_exception)
-        return msg
-
-
 class Cassandra(object):
-    """Cassandra Backend Connector
+    """Cassandra Back-end Connector
 
     Taxonomies and TagPacks can be ingested into Apache Cassandra
     for further processing within GraphSense.
@@ -58,7 +42,7 @@ class Cassandra(object):
     def setup_keyspace(self, keyspace):
         """Setup keyspace and tables"""
         if not self.session:
-            raise StorageError("Session not availble. Call connect() first")
+            raise StorageError("Session not available. Call connect() first")
 
         schema = pkg_resources.read_text(db, SCHEMA_FILE)
 
@@ -100,15 +84,6 @@ class Cassandra(object):
         except Exception as e:
             raise StorageError("Taxonomy insertion error", e)
 
-    def _add_normalized_label(self, tag_json):
-        """Enrich JSON Tag representation by normalized labels"""
-        # Alphanumeric and lowercase only
-        d = json.loads(tag_json)
-        label = d['label']
-        d['label_norm'] = LABEL_NORM_PATTERN.sub('', label).lower()
-        d['label_norm_prefix'] = d['label_norm'][:3]
-        return json.dumps(d)
-
     def insert_tagpack(self, tagpack, keyspace, concurrency):
         """Insert a tagpack into a given keyspace"""
         if not self.session:
@@ -116,25 +91,30 @@ class Cassandra(object):
         try:
             self.session.set_keyspace(keyspace)
 
-            q = f"INSERT INTO tagpack_by_uri JSON '{tagpack.to_json()}'"
+            tagpack_json = tagpack.to_json()
+            # tagpack_json = Cassandra.fields_to_timestamp(tagpack_json)
+
+            q = f"INSERT INTO tagpack_by_uri JSON '{tagpack_json}'"
             self.session.execute(q)
 
-            stmt_1 = self.session.prepare("INSERT INTO tag_by_address JSON ?")
-            stmt_2 = self.session.prepare("INSERT INTO tag_by_category JSON ?")
-            stmt_3 = self.session.prepare("INSERT INTO tag_by_label JSON ?")
+            stmt_addr = self.session.prepare(
+                "INSERT INTO address_tag_by_address JSON ?")
+
+            stmt_entity = self.session.prepare(
+                "INSERT INTO entity_tag_by_id JSON ?")
 
             statements_and_params = []
+
             for tag in tagpack.tags:
                 # prepare statements for table tag_by_address
                 params = (tag.to_json(), )
-                statements_and_params.append((stmt_1, params))
-                # prepare statements for table tag_by_category
-                if 'category' in tag.fields.keys():
-                    statements_and_params.append((stmt_2, params))
-                # prepare statements for table tag_by_label
-                tag_w_norm_labels = self._add_normalized_label(tag.to_json())
-                params = (tag_w_norm_labels, )
-                statements_and_params.append((stmt_3, params))
+                if isinstance(tag, AddressTag):
+                    statements_and_params.append((stmt_addr, params))
+                elif isinstance(tag, EntityTag):
+                    statements_and_params.append((stmt_entity, params))
+                else:
+                    raise StorageError(
+                        "Invalid tag: must be either Address or EntiyTag")
 
             results = execute_concurrent(self.session, statements_and_params,
                                          concurrency=concurrency,
@@ -148,5 +128,5 @@ class Cassandra(object):
             raise StorageError("TagPack insertion error", e)
 
     def close(self):
-        """Closes the cassandra cluster connection"""
+        """Closes the Apache Cassandra cluster connection"""
         self.cluster.shutdown()